refactor(button event handler): place a button handler logic into a single function
